requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.0
azure-storage-blob==12.18.2
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
Deprecated==1.2.14
dnspython==2.4.2
emoji==2.8.0
exceptiongroup==1.1.3
fastapi==0.103.2
filetype==1.2.0
Flask==2.3.3
frozenlist==1.4.0
future==0.18.3
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.304
langdetect==1.0.9
langsmith==0.0.41
line-bot-sdk==3.5.0
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.4
openai==0.28.1
packaging==23.2
pinecone-client==2.2.4
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.10.18
urllib3==2.0.6
uvicorn==0.23.2
Werkzeug==3.0.0
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
